[ci skip] Reorganize context
Split up logic into class, function and field.
